+++ src/System.Net.Primitives/tests/PalTests/Fakes/NetEventSource.cs
@@ -8,6 +8,7 @@ namespace System.Net
     {
         public static void Enter(object thisOrContextObject, object arg1 = null, object arg2 = null, object arg3 = null) { }
         public static void Fail(object thisOrContextObject, object arg) { }
+        public static void Info(object thisOrContextObject, object arg) { }
         public static bool IsEnabled => false;
     }
 }
